# public/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "NPI Website API",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Development Server"
    }
  ],
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT Authorization header using the Bearer scheme."
      }
    },
...
  },
  "paths": {
...
    "/api/v1/factory/nonconformance/close": {
      "post": {
        "summary": "Close nonconformances for multiple things",
        "description": "Closes all open nonconformances for specified things with \"NO PROBLEM FOUND\" disposition",
        "tags": ["Factory"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["thingNames", "token"],
                "properties": {
                  "thingNames": {
                    "type": "string",
                    "description": "Comma-separated list of thing names",
                    "example": "THING001,THING002,THING003"
                  },
                  "token": {
                    "type": "string",
                    "description": "JWT authentication token"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Nonconformances closed successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "NCs closed successfully"
                    },
                    "results": {
                      "type": "object",
                      "properties": {
                        "successful": {
                          "type": "array",
                          "items": {
                            "type": "string"
                          },
                          "description": "List of successfully processed thing names"
                        },
                        "failed": {
                          "type": "array",
                          "items": {
                            "type": "object",
                            "properties": {
                              "thingName": {
                                "type": "string"
                              },
                              "error": {
                                "type": "string"
                              }
                            }
                          },
                          "description": "List of failed operations with error details"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad request or validation error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    },
                    "results": {
                      "type": "object",
                      "properties": {
                        "failed": {
                          "type": "array",
                          "items": {
                            "type": "object",
                            "properties": {
                              "error": {
                                "type": "string"
                              }
                            }
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/factory/process/get-process/{slug}": {
      "get": {
        "summary": "Get non-production flows for a process",
        "description": "Retrieves all non-production flows for a specified process name",
        "tags": ["Factory"],
        "parameters": [
          {
            "in": "path",
            "name": "slug",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Process name to get flows for",
            "example": "PROCESS_001"
          }
        ],
        "responses": {
          "200": {
            "description": "Non-production flows retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Successfully fetched flows"
                    },
                    "results": {
                      "type": "object",
                      "properties": {
                        "flows": {
                          "type": "array",
                          "items": {
                            "type": "object"
                          },
                          "description": "Array of non-production flows"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad request or validation error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    },
                    "results": {
                      "type": "object",
                      "properties": {
                        "failed": {
                          "type": "array",
                          "items": {
                            "type": "object",
                            "properties": {
                              "error": {
                                "type": "string"
                              }
                            }
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/factory/thing/change-part-number": {
      "post": {
        "summary": "Change part number for one or more things",
        "description": "Changes the part number of specified thing(s) to a given part number",
        "tags": ["Factory"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["partNumber", "thingNames", "token", "username"],
                "properties": {
                  "partNumber": {
                    "type": "string",
                    "description": "Target part number in format 0000000-00-X",
                    "example": "1234567-00-A"
                  },
                  "thingNames": {
                    "type": "string",
                    "description": "Comma-separated list of thing names",
                    "example": "THING001,THING002,THING003"
                  },
                  "token": {
                    "type": "string",
                    "description": "JWT authentication token"
                  },
                  "username": {
                    "type": "string",
                    "description": "Username of the actor performing the action",
                    "example": "user@tesla.com"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Part number changes completed successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    },
                    "results": {
                      "type": "object",
                      "properties": {
                        "successful": {
                          "type": "array",
                          "items": {
                            "type": "string"
                          }
                        },
                        "failed": {
                          "type": "array",
                          "items": {
                            "type": "object",
                            "properties": {
                              "thingName": {
                                "type": "string"
                              },
                              "error": {
                                "type": "string"
                              }
                            }
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Failed to change part numbers",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    },
                    "results": {
                      "type": "object",
                      "properties": {
                        "failed": {
                          "type": "array",
                          "items": {
                            "type": "object",
                            "properties": {
                              "thingName": {
                                "type": "string"
                              },
                              "error": {
                                "type": "string"
                              }
                            }
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
...
}